+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.2.0
 
-Refactors script descriptors module in order to make invalid descriptors unrepresentable.
+* Refactors script descriptors module in order to make invalid descriptors unrepresentable.
+* Adds utility functions to compile script descriptors and produce addresses from output descriptors.
 
 ## 0.1.0